# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Excelsior Deckbuilder API v1
  description: |
    External HTTP API for the Excelsior Overpower card database and deck builder.

    Stability: this is a stub spec. Paths and payloads are authoritative from
    [`API_V1.md`](../API_V1.md); this document is maintained alongside it so
    external clients can generate SDKs. Any new `/api/v1` route must add an
    entry here, an error-catalog entry, and a changelog entry in the same PR.

    Contract highlights:

      - Responses use the v1 envelope `{ data, meta, errors }`.
      - Authentication is `Bearer <accessToken>` JWT (see `/auth/login`).
      - Refresh tokens rotate on `/auth/refresh`; reuse revokes the family.
      - Rate limits are per-user (or per-IP for unauthenticated calls). See
        `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.
      - Catalog endpoints are edge-cached with `Cache-Control` + `ETag`; clients
        SHOULD send `If-None-Match` to get `304 Not Modified`.
      - Images are served from `APP_CDN_BASE` over HTTPS. See
        [`API_V1_IMAGE_CONTRACT.md`](current/API_V1_IMAGE_CONTRACT.md).
  version: 1.0.0
  contact:
    name: Excelsior
    url: https://excelsior.cards
servers:
  - url: https://excelsior.cards
    description: Production
  - url: http://localhost:8085
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Envelope:
      type: object
      properties:
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/Meta'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorBody'
      required: [data, meta, errors]
    Meta:
      type: object
      properties:
        requestId:
          type: string
        catalogDataVersion:
          type: integer
          minimum: 0
          description: Monotonic version for all catalog/DBV responses. Bumps on re-ingest.
        catalogLastUpdated:
          type: string
          format: date-time
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable error code. See API_V1_ERROR_CATALOG.md.
        message:
          type: string
        field:
          type: string
          description: Present for VALIDATION_ERROR.
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Username or email.
        password:
          type: string
    LoginSuccess:
      type: object
      required: [accessToken, expiresInSeconds, user]
      properties:
        accessToken:
          type: string
        expiresInSeconds:
          type: integer
        refreshToken:
          type: string
        refreshExpiresInSeconds:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, USER, GUEST]
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in and obtain access + refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginSuccess'
        '401':
          description: Invalid credentials.
        '429':
          description: Rate limited.
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the refresh token and receive a new access+refresh pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Rotation success.
        '401':
          description: Refresh token invalid, expired, or revoked.
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the provided refresh token (if any).
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout acknowledged.
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user.
      responses:
        '200':
          description: User envelope.
        '401':
          description: Not authenticated.
...